+++ Tests/IGListDiffTests.m
@@ -366,4 +366,66 @@ static NSArray *sorted(NSArray *arr) {
     XCTAssertEqualObjects([result newIndexPathForIdentifier:@9], [NSIndexPath indexPathForItem:1 inSection:1]);
 }
 
+- (void)test_whenDiffing_withBatchUpdateResult_thatIndexesMatch {
+    NSArray *o = @[
+                   genTestObject(@1, @1),
+                   genTestObject(@2, @1),
+                   genTestObject(@3, @1),
+                   genTestObject(@4, @1),
+                   genTestObject(@5, @1),
+                   genTestObject(@6, @1),
+                   ];
+    NSArray *n = @[
+                   // deleted
+                   genTestObject(@2, @2), // updated
+                   genTestObject(@5, @1), // moved
+                   genTestObject(@4, @1),
+                   genTestObject(@7, @1), // inserted
+                   genTestObject(@6, @2), // updated
+                   genTestObject(@3, @2), // moved+updated
+                   ];
+    IGListIndexSetResult *result = [IGListDiff(o, n, IGListDiffEquality) resultForBatchUpdates];
+    XCTAssertEqual(result.updates.count, 0);
+    NSArray *expectedMoves = @[ [[IGListMoveIndex alloc] initWithFrom:4 to:1] ];
+    XCTAssertEqualObjects(result.moves, expectedMoves);
+    NSMutableIndexSet *expectedDeletes = [NSMutableIndexSet indexSetWithIndex:0];
+    [expectedDeletes addIndex:1];
+    [expectedDeletes addIndex:2];
+    [expectedDeletes addIndex:5];
+    XCTAssertEqualObjects(result.deletes, expectedDeletes);
+    NSMutableIndexSet *expectedInserts = [NSMutableIndexSet indexSetWithIndex:0];
+    [expectedInserts addIndex:3];
+    [expectedInserts addIndex:4];
+    [expectedInserts addIndex:5];
+    XCTAssertEqualObjects(result.inserts, expectedInserts);
+}
+
+- (void)test_whenDiffing_withBatchUpdateResult_thatIndexPathsMatch {
+    NSArray *o = @[
+                   genTestObject(@1, @1),
+                   genTestObject(@2, @1),
+                   genTestObject(@3, @1),
+                   genTestObject(@4, @1),
+                   genTestObject(@5, @1),
+                   genTestObject(@6, @1),
+                   ];
+    NSArray *n = @[
+                   // deleted
+                   genTestObject(@2, @2), // updated
+                   genTestObject(@5, @1), // moved
+                   genTestObject(@4, @1),
+                   genTestObject(@7, @1), // inserted
+                   genTestObject(@6, @2), // updated
+                   genTestObject(@3, @2), // moved+updated
+                   ];
+    IGListIndexPathResult *result = [IGListDiffPaths(0, 1, o, n, IGListDiffEquality) resultForBatchUpdates];
+    XCTAssertEqual(result.updates.count, 0);
+    NSArray *expectedMoves = @[ [[IGListMoveIndexPath alloc] initWithFrom:genIndexPath(4, 0) to:genIndexPath(1, 1)] ];
+    XCTAssertEqualObjects(result.moves, expectedMoves);
+    NSArray *expectedDeletes = @[genIndexPath(0, 0), genIndexPath(1, 0), genIndexPath(2, 0), genIndexPath(5, 0)];
+    XCTAssertEqualObjects(sorted(result.deletes), expectedDeletes);
+    NSArray *expectedInserts = @[genIndexPath(0, 1), genIndexPath(3, 1), genIndexPath(4, 1), genIndexPath(5, 1)];
+    XCTAssertEqualObjects(sorted(result.inserts), expectedInserts);
+}
+
 @end