+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The changelog for `IGListKit`. Also see the [releases](https://github.com/instag
 
 This release closes the [2.0.0 milestone](https://github.com/Instagram/IGListKit/milestone/1?closed=1).
 
+### Breaking Changes
+
+- Diff result method `-resultWithUpdatedMovesAsDeleteInserts` removed and replaced with `-resultForBatchUpdates`
+
 ### Enhancements
 
 - Added support for cells created from nibs. [Sven Bacia](https://github.com/svenbacia) [(#56)](https://github.com/Instagram/IGListKit/pull/56)